directives in separate files are combined into single file (main-validator) and services file modified also.
