--- Well/src/app/directive/email-validator.directive.spec.ts
@@ -1,8 +1,0 @@
-import { EmailValidatorDirective } from './email-validator.directive';
-
-describe('EmailValidatorDirective', () => {
-  it('should create an instance', () => {
-    const directive = new EmailValidatorDirective();
-    expect(directive).toBeTruthy();
-  });
-});
