--- Well/src/app/directive/name-validator.directive.spec.ts
@@ -1,8 +1,0 @@
-import { NameValidatorDirective } from './name-validator.directive';
-
-describe('NameValidatorDirective', () => {
-  it('should create an instance', () => {
-    const directive = new NameValidatorDirective();
-    expect(directive).toBeTruthy();
-  });
-});
