--- Well/src/app/directive/pass-validator.directive.spec.ts
@@ -1,8 +1,0 @@
-import { PassValidatorDirective } from './pass-validator.directive';
-
-describe('PassValidatorDirective', () => {
-  it('should create an instance', () => {
-    const directive = new PassValidatorDirective();
-    expect(directive).toBeTruthy();
-  });
-});
