--- Well/src/app/directive/phone-validator.directive.spec.ts
@@ -1,8 +1,0 @@
-import { PhoneValidatorDirective } from './phone-validator.directive';
-
-describe('PhoneValidatorDirective', () => {
-  it('should create an instance', () => {
-    const directive = new PhoneValidatorDirective();
-    expect(directive).toBeTruthy();
-  });
-});
