+++ Well/src/app/profile/profile.component.html
@@ -5,13 +5,13 @@
     <h2> You are Logged In!</h2><span id="logout" (click)="userLogOut()">Logout</span>
 
     <div class="image">
-      <img src="assets/photo.png" height="200" width="200"/>
+      <img src="assets/images/photo.png" height="200" width="200"/>
     </div>
 
     <div class="data">
-        <h3 class="key"> Name  : <span class="value">Gursewak</span></h3>
-        <h3 class="key"> Phone : <span class="value"> Gursewak</span></h3>
-        <h3 class="key"> Email : <span class="value"> Gursewak</span></h3>
+        <h3 class="key"> Name  : <span class="value"> {{name}} </span></h3>
+        <h3 class="key"> Phone : <span class="value"> {{phone}} </span></h3>
+        <h3 class="key"> Email : <span class="value"> {{email}} </span></h3>
     </div>
 
   </div>
